+++ modules/antivirus/eset/nod32.py
@@ -15,13 +15,14 @@
 
 import logging
 import re
+import os
 
-from ..base import Antivirus
+from modules.antivirus.base import Antivirus
 
 log = logging.getLogger(__name__)
 
 
-class EsetNod32(Antivirus):
+class nod32(Antivirus):
 
     # ==================================
     #  Constructor and destructor stuff
@@ -29,16 +30,26 @@ class EsetNod32(Antivirus):
 
     def __init__(self, *args, **kwargs):
         # class super class constructor
-        super(EsetNod32, self).__init__(*args, **kwargs)
+        super(nod32, self).__init__(*args, **kwargs)
         # set default antivirus information
-        self._name = "ESET NOD32 Antivirus Business Edition for Linux Desktop"
-        # Modify retun codes (see --help for details)
-        self._scan_retcodes[self.ScanResult.INFECTED] = lambda x: x in [1, 50]
+        if self._is_windows:
+            self._name = "ESET8 Anti-Virus for Windows"
+        else:
+            self._name = "ESET NOD32 Antivirus Business Edition for Linux Desktop"
         # scan tool variables
-        self._scan_args = (
+        if self._is_windows:
+            self._scan_args = (
+            "/unwanted "  # scan command
+            "/log-rewrite"   # report only
+			)
+        else:
+             # scan tool variables
+            self._scan_args = (
             "--clean-mode=NONE "  # do not remove infected files
             "--no-log-all"        # do not log clean files
-        )
+			)
+        # TODO: check for retcodes in WINDOWS & Linux
+        self._scan_retcodes[self.ScanResult.INFECTED] = lambda x: x in [1, 50]
         self._scan_patterns = [
             re.compile(r'name="(?P<file>.*)", threat="(?P<name>.*)", '
                        r'action=.*', re.IGNORECASE | re.MULTILINE)
@@ -50,26 +61,40 @@ class EsetNod32(Antivirus):
 
     def get_version(self):
         """return the version of the antivirus"""
-        result = None
-        if self.scan_path:
+        result = None	
+        if self._is_windows:
+            cmd = self.build_cmd(self.scan_path, '/version')
+            retcode, stdout, stderr = self.run_cmd(cmd)
+
+   
+        else:
             cmd = self.build_cmd(self.scan_path, '--version')
             retcode, stdout, stderr = self.run_cmd(cmd)
-            if not retcode:
-                matches = re.search(r'(?P<version>\d+(\.\d+)+)',
+			
+        if not retcode:
+            matches = re.search(r'(?P<version>\d+(\.\d+)+)',
                                     stdout,
                                     re.IGNORECASE)
-                if matches:
-                    result = matches.group('version').strip()
+        if matches:
+         result = matches.group('version').strip()
         return result
 
     def get_database(self):
         """return list of files in the database"""
-        search_paths = [
-            '/var/opt/eset/esets/lib/',
-        ]
-        database_patterns = [
+        if self._is_windows:
+            search_paths = map(lambda x:
+        "{path}/ESET/*/ESET Smart Security".format(path=x),
+            [os.environ.get('PROGRAMDATA', '')])
+            database_patterns = [
             '*.dat',  # determined using strace on linux
-        ]
+                    ]
+        else:
+           search_paths = [
+            '/var/opt/eset/esets/lib/',
+			]
+           database_patterns = [
+            '*.dat',  # determined using strace on linux
+			]
         results = []
         for pattern in database_patterns:
             result = self.locate(pattern, search_paths, syspath=False)
@@ -78,5 +103,12 @@ class EsetNod32(Antivirus):
 
     def get_scan_path(self):
         """return the full path of the scan tool"""
-        paths = self.locate("esets_scan", "/opt/eset/esets/sbin/")
+        if self._is_windows:
+            scan_bin = "ecls.exe"
+            scan_paths = map(lambda x: "{path}/ESET/*".format(path=x),
+                         [os.environ.get('PROGRAMFILES', ''),
+                          os.environ.get('PROGRAMFILES(X86)', '')])
+            paths = self.locate(scan_bin, scan_paths)
+        else:
+             paths = self.locate("esets_scan", "/opt/eset/esets/sbin/")
         return paths[0] if paths else None
